refactor(interaction-controller): move setCamerView method
